clients/fast-api/requirements.txt
boto3
fastapi
pyjwt
python-multipart
uvicorn